Correct content of components
